--- functions.py
@@ -1,21 +1,0 @@
-#!/usr/bin/env python
-# -*- coding:utf-8 -*-
-
-"""The package's functions
-
-Slight Fimulator - Flight simulator in Python
-Copyright (C) 2017 Hao Tian and Adrien Hopkins
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
-
-# Installs Python 3 division and print behaviour
-from __future__ import division, print_function
