--- unofficial_utilities_for_pygame/README.md
@@ -1,18 +1,0 @@
-# Unofficial Utilities for Pygame
-An unofficial module that makes pygame coding easier.
-
-## Dependencies
-Requires Python 2 or 3, Pygame
-
-This command installs them on GNU/Linux:
-
-Python 2: `sudo apt-get update && sudo apt-get install python python-pip && sudo pip install pygame`  
-Python 3: `sudo apt-get update && sudo apt-get install python3 python3-pip && sudo pip3 install pygame`  
-
-# How to Use
-This is not a standalone script. You must import it from within python.
-To make this module importable, you must rename it to something with only letters, numbers and underscores (the first character may not be a number)
-
-Within your program:
-
-`import (module name)`
